// src/app/shared/types/api-types/stock-information-dto-interface.ts
export interface StockInformationDTO {
  pn: string;
  designation: string;
  siteId: string;
  siteText: string;
  warehouseId: string;
  warehouseText: string;
  storageBinId: string;
  storageBinText: string;
  stockTypeId: string;
  stockTypeText: string;
  valuationGroupId: string;
  valuationGroupText: string;
  stockSn: string;
  stockQuantity: string;
  stockBatchNumber: string;
  stockReceiptDate: Date;
  stockReceiptQuantity: string;
  stockIssueDate: Date;
  stockIssueQuantity: string;
  stockIssueWorkOrderWoId: string;
  stockIssueWorkOrderWoCode: string;
  stockReceiptWorkOrderWoId: string;
  stockReceiptWorkOrderWoCode: string;
  stockQuantityDtos: StockInformationDTO[];
  stockIssueDtos: StockInformationDTO[];
  issueDateStr: string;
  quantityDateStr: string;
  operationalStatus?: string;
  interchangeabilityText?: string;
  areaId?: string;
  areaText?: string;
  manufacturingBatchNumber?: string;
  stockMvtUnit?: string;
}